Refactor lincs client further.
